requirements.txt
streamlit
langchain
langchain_core
langchain_community
python-dotenv
langchain-groq
plotly
pandas
crewai
pydantic
wikipedia
yfinance
unstructured
sec-api
psutil